README.md
Description:

Simple Rails app used to search for flights between specified airports on a certain date. Can create a booking of flights, which gathers passenger information. All data is made up.

To run locally: 

Clone Repo

In project directory that you cloned into:
    rails db:migrate #setup the database
    rails db:seed #to seed the dummy data so you have something to look at when you run it. 
    rails s #by default this will run in localhost:3000, so go there to play


This project was created as part of the Odin Project, and the specifications can be found here: https://www.theodinproject.com/paths/full-stack-ruby-on-rails/courses/ruby-on-rails/lessons/flight-booker